Implemented basic drawing for blocks.
All blocks are defined by a list of vertices,
where each vertex is connected by an edge to
the vertices before and after it in the list.
The first and last vertices in the list are
connected. Use Block::add_vertex to add vertices.
There is currently no way to modify an existing
block's vertices except by adding more, but it
should be trivial to add such methods.

Block.cpp contains function definitions of
Block methods, particularly Block::init_render_cache
which relies upon Block::triangulate, which relies
upon PolygonDCEL.

PolygonDCEL is an implementation of a doubly
connected edge list specialized for the use-case
of triangulation of a closed polygon.
More on DCELs: https://www2.cs.sfu.ca/~binay/813.2011/DCEL.pdf

The Block::triangulate() function uses an algorithm
described in the short paper linked below to
divide the block into triangles. This is primarily
for drawing purposes as OpenGL primarily handles
triangles. It may be useful for collision detection.
https://www.cs.umd.edu/class/spring2020/cmsc754/Lects/lect05-triangulate.pdf