Windows VSCode compatibility changes.
The biggest side effect of this change is that
shaders are now copied into source files during
CMake generation.

This was primarily implemented
because it is difficult to place the shaders directory
in the binary directory when using Visual Studio
for compilation, which is largely used in parallel
with VSCode on Windows.