Fixed circular header dependency and restructured fps const var distribution
It looks like there's a Constants.h header which hasn't
been pushed to the repo, so I was unable to build the
project. Regardless, it is probably a better idea to
store the fps as a field of the Screen class, as global
variables seem to be generally frowned upon.