Implemented normal calculation and partial solving
I need to make some brief finishing touches to CollisionGroup::solve and
review my usage of global vs relative coordinates (I guarantee it is not
consistent, but it needs to be). Testing should begin at ~3-4 pm 5/14,
and will not initially depend on the separate collision detection code
(that would make testing very complicated).